Renamed WritePlan -> SectionsContent & WritePlanLayout -> WritePlan
