+++ src/App.js
@@ -1,16 +1,15 @@
 import "./App.css";
-import Redaction from "./components/WritePlan/WritePlan";
 import Form from "./components/Forms/Form";
 import PlanCreation from "./components/PlanCreation/PlanCreation";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import WritePlanLayout from "./components/WritePlan/WritePlanLayout";
+import WritePlan from "./components/WritePlan/WritePlan";
 import GeneralInfo from "./components/GeneralInfo/GeneralInfo";
 
 function App() {
   return (
     <Router>
       <Routes>
-        <Route exact path="/redaction" element={<WritePlanLayout readonly={false} />} />
+        <Route exact path="/redaction" element={<WritePlan readonly={false} />} />
         <Route exact path="/form" element={<Form schemaId={"a"} readonly={true} />} />
         <Route exact path="/plan" element={<PlanCreation />} />
         <Route exact path="/" element={<GeneralInfo readonly={false} />} />
